+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # changelog
 
 ## [unreleased]
+### fixed
+* added `tokio-util` and `signal-hook-registry` to `smolbar --license`
 
 ## [0.8.0] - 2023-04-22
 ### added